+++ app/controllers/frontend/daggerheart/characters/rest_controller.rb
@@ -13,7 +13,7 @@ module Frontend
         def create
           case change_energy.call(rest_options.merge({ character: @character }))
           in { errors: errors, errors_list: errors_list } then unprocessable_response(errors, errors_list)
-          else only_head_response
+          in { rolls: rolls } then render json: { rolls: rolls }, status: :ok
           end
         end